+++ apps/web/src/components/event-calendar/event-context-menu.tsx
@@ -18,7 +18,7 @@ import { KeyboardShortcut } from "@/components/ui/keyboard-shortcut";
 import { Tooltip, TooltipContent } from "@/components/ui/tooltip";
 import { useTRPC } from "@/lib/trpc/client";
 import { cn } from "@/lib/utils";
-import { Action } from "./event-calendar/hooks/use-event-operations";
+import { Action } from "./hooks/use-event-operations";
 
 function CalendarRadioItem({
   className,
